feat: lazy load webgl canvas
