+++ src/components/ThreeDBackground/ThreeDBackground.module.scss
@@ -6,12 +6,27 @@
   height: 100vh;
   z-index: -2;
   pointer-events: none;
-  
+
   canvas {
     width: 100% !important;
     height: 100% !important;
     background: transparent;
   }
+}
+
+.noiseFallback {
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  pointer-events: none;
+  z-index: -2;
+  background-image: radial-gradient(circle, rgba(255, 255, 255, 0.05) 1px, transparent 1px),
+    radial-gradient(circle, rgba(255, 255, 255, 0.05) 1px, transparent 1px);
+  background-size: 3px 3px;
+  background-position: 0 0, 1.5px 1.5px;
+  opacity: 0.2;
 }
 
 .overlay {
